Added data related modules and classes
- Data::DataSource in
	data/data_source.rb
- Data::Google::DatastoreSource in
	data/google/datastore_source.rb
- Data::DataContext in
	data/data_context.rb

The Data::DataSource is the beginning of an abstraction for data
sources. There will be child classes which implement specific providers.
For example the Data::Google::DatastoreSource will be used to interact
with the google cloud datastore platform. The idea behind the
Data::DataSource classes are to allow easy switching between providers
for the Data::DataContext classes which will use the Data::DataSource
classes in turn.

Data::DataContext will function as a base class with a purpose
to interact with the Data::DataSource class. It will be inherited
from by implementations with a more focused purpose. For example in
future there will be a GameTokenDataContext class.